+++ requests/packages/urllib3/__init__.py
@@ -10,26 +10,20 @@ urllib3 - Thread-safe connection pooling and re-using.
 
 __author__ = 'Andrey Petrov (andrey.petrov@shazow.net)'
 __license__ = 'MIT'
-__version__ = '1.2.2'
+__version__ = '1.3'
 
 
 from .connectionpool import (
     HTTPConnectionPool,
     HTTPSConnectionPool,
-    connection_from_url,
-    get_host,
-    make_headers)
+    connection_from_url
+)
 
-
-from .exceptions import (
-    HTTPError,
-    MaxRetryError,
-    SSLError,
-    TimeoutError)
-
+from . import exceptions
+from .filepost import encode_multipart_formdata
 from .poolmanager import PoolManager, ProxyManager, proxy_from_url
 from .response import HTTPResponse
-from .filepost import encode_multipart_formdata
+from .util import make_headers, get_host
 
 
 # Set default logging handler to avoid "No handler found" warnings.